auth and manage listings

+++ app/Http/Controllers/ListingController.php
@@ -42,6 +42,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -49,6 +51,12 @@ class ListingController extends Controller
 
     // show edit form
     public function edit(Listing $listing) {
+
+        // make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('listings.edit', ['listing' => $listing]);
     }
 
@@ -75,8 +83,20 @@ class ListingController extends Controller
 
     // delete listing
     public function destroy(Listing $listing) {
+
+        // make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    // manage listings
+    public function manage() {
+
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
