+++ routes/web.php
@@ -1,8 +1,9 @@
 <?php
 
-use App\Http\Controllers\ListingController;
-use Illuminate\Support\Facades\Route;
 use App\Models\Listing;
+use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
 /*
 |--------------------------------------------------------------------------
@@ -28,19 +29,38 @@ use App\Models\Listing;
 Route::get('/', [ListingController::class, 'index']);
 
 // show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // store listing data
 Route::post('/listings', [ListingController::class, 'store']);
 
 // show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // edit submit to update
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+
+// manage listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // single listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
+
+
+// show register create form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+// create new user
+Route::post('/users', [UserController::class, 'store']);
+
+// logout user
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
